Second try cleaning up progress
